main:
In components/SpotlightCard.css, changed spotlight-color attribute.
In components/Squares.css, added position to fixed and removed display block attribute.
In ModularAlternateReverseTimeline.jsx, added last connector, as to allow scrolling down to the point
where the last card is not behind the dock.
Added ModularLinearTimeline.jsx for the mobile version of the timeline.
Added ResponsiveModularTimeline.jsx, to bifurcate the rendering of the timeline between
ModularAlternateReverseTimeline and ModularLinearTimeline.
Added ProjectsAPI.jsx and SkillsAPI.jsx.