main:
In MyCardSwap.css, changed transform attribute in media queries and not placing swap cards in a better
position in the container.
In PillNav.css, added background color with gradient.
In HeaderWindowContainer.css, commented out margin and padding.
In HeaderWindowContainerPage.css, added box-shadow.
In ShuffleProjectPageTitle.css, added css classes for each word in the title of Projects page.