main:
In PillNav.css, removed padding since it is now a child of app-inner-wrap container.
In OrientationBasedOnIndexTimelineSpotlightCard.css, added hover behaviour with drop-shadow
attribute for the right-timeline-spotlight-card and left-timeline-spotlight-card. Added
it only for their direct child .card-spotlight, since before the whole container (bigger than the card)
would enable the hover behaviour.
In OrientationBasedOnIndexTimelineSpotlightCard.jsx and RightOrientedTimelineSpotlightCard.jsx, removed
assignment of spotlightColor since it is defined in the css file.
In HeaderWindowContainer.css, added margin (left-right) in media query for 480px.
In App.css added app-outer-wrap and app-inner-wrap with media query so that for smaller screens
there is padding (left-right).
In App.jsx, encapsulated PillNav, Routes and Dock in app-inner-wrap, and MyDotGrid together with
app-inner-wrap div in the app-outer-wrap.