main:
In MyCardSwap.css, changed changed max width of media query to be alligned with that of the Project container.
In PillNav.css, changed styling of mobile menu items, to be consistent with the general pallette of the project.
In Home.jsx, added rotating text component (working, but needs tweaking for smoother transitions).
Created Home.css.
Created component RotatingText (from Reactbits.com).