main:
In App.jsx, removed Dock component and moved to Bio.jsx so that it can be under the bio-page-container css class. Should be moved
to the rest of the pages as well.
In ModularLinearTimeline.jsx, the timeline is now unified, that is the spotlight content cards
are positioned in sequence to the timeline components.
In OrientationBasedOnIndexTimelineSpotlightCard.css, added linear-timeline-spotlight-card for use in mobile.
Renamed RightOrientedTimelineSpotlightCard to LinearTimelineSpotlightCard.jsx.
Added ModularAlternateResponsiveTimelineContainer.css containing the classes for the timeline wrapper.